Added Files
I changed the version numeration of the files. Completed first phase of the development. Basic interface to interact. Added a few commands. Changing mac address has to be implemented as well as all the fake websites, apache configuration, install.sh script, the evil twin attack and the scanning command. The apache configuration has to be able to work with php using hostapd and the dns mask.